Changed the data types to the right sizes (100% correct)

+++ src/c/header/stdint.h
@@ -5,13 +5,13 @@
 
 /* exact-width integer types */
 typedef signed char             int8_t;
-typedef short int               int16_t;
-typedef int                     int32_t;
-typedef long long               int64_t;
+typedef signed short int        int16_t;
+typedef signed long int         int32_t;
+typedef signed long long int    int64_t;
 typedef unsigned char           uint8_t;
 typedef unsigned short int      uint16_t;
-typedef unsigned int            uint32_t;
-typedef unsigned long long      uint64_t;
+typedef unsigned long int       uint32_t;
+typedef unsigned long long int  uint64_t;
 
 /* minimum-width integer types */
 typedef int8_t                  int_least8_t;
